instructions/cleanup

+++ scripts/copy-files.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/bash
+#!/usr/bin/bash -e
 
 mc alias set local https://minio.la.utexas.edu "$S3_ACCESS_KEY" "$S3_SECRET_KEY"
 
@@ -12,7 +12,13 @@ while IFS= read -r interviewsectionid; do
     fi
 
     echo "copying $interviewsectionid"
-    mc cp --recursive local/speech-islands/interviews/$prefix/$interviewsectionid.eaf /corpusdata/TGDP/$prefix
-    mc cp --recursive local/speech-islands/sound_files/$prefix/$interviewsectionid.mp3 /corpusdata/TGDP/$prefix
-    mc cp --recursive local/speech-islands/sound_files/$prefix/$interviewsectionid.wav /corpusdata/TGDP/$prefix
+    if [[ ! -f "/corpusdata/TGDP/$prefix/$interviewsectionid.eaf" ]]; then
+      mc cp --recursive local/speech-islands/interviews/$prefix/$interviewsectionid.eaf /corpusdata/TGDP/$prefix
+    fi
+    if [[ ! -f "/corpusdata/TGDP/$prefix/$interviewsectionid.mp3" ]]; then
+      mc cp --recursive local/speech-islands/sound_files/$prefix/$interviewsectionid.mp3 /corpusdata/TGDP/$prefix
+    fi
+    if [[ ! -f "/corpusdata/TGDP/$prefix/$interviewsectionid.wav" ]]; then
+      mc cp --recursive local/speech-islands/sound_files/$prefix/$interviewsectionid.wav /corpusdata/TGDP/$prefix
+    fi
 done < "public-sections.txt"
